One more fix for #838
If there's no Giants turn radius for a towed implement,
use 6 m as a default (as opposed to the turn radius of the
tractor, which can be very small, causing pinching).

Also, for now, disabled tight turn offset for tracked and
articulated axis vehicles. This probably isn't the best solution,
maybe they need a lower offset or a longer look ahead distance.