Push Release 7.0.0.27
Important:
7.0.0.26 had a bug which resulted in incorrectly saved courses (sorry, our bad). This bug was most likely contributing to issues like #883, #890 and #890.
There is no easy way to tell if a saved course is corrupt or not. Therefore, we recommend removing all saved courses and regenerate them with 7.0.0.27. The temporary courses saved with the vehicles are also affected, so the safest bet is to remove these from the vehicles after the game loads and regenerate them as needed. Our sincere apologies for the inconvenience.

- Fixes course save bug
- Highlights the last clicked point, while drawing a custom course on the map.
- Mower works everywhere, for #826
- Disables hud elements mouse event, if they are invisble or disabled.
- Use pathfinder to drive to fieldwork start #879
- Increased field margin to -5 #888
- Finding bales on custom fields #532
- Only generate offset course when needed #883 #893 #890
- Ignore fruit when driving to start field work