Fix for 180 turn offset multitool
When there is connecting track (which overlaps a part
of the headland) we sometimes select the wrong waypoint
to start/end the section of the turn following the headland.

Made sure no points on the overlapping headland part are used
in a wide turn.

Note that this does not fix the missing connecting track labels
on the left/right offset courses.